app/requirements/requirements.txt
djongo
django==3.0.5
djangorestframework
markdown      
django-filter  
gunicorn
django-filter
whitenoise
requests